[content] Mom and her neighbor speeches are done.
Known bug: the cases where mom is supposed to stand at the different hours of the day make a sound when you press A while facing it (even though it does nothing else). It is because the bg_event still exists there. I can't get rid of it at the moment, but it is a minor bug. Happens with the neighbor as well. I need those bg_events so you can talk face to face with mom and her neighbor before the player learns about social distancing.