[content] road to goldenrod game corner done.
Now the interior part needs to be improved, with a quick onboarding that gives the coin box, 50 coins and explains to use the Abra to teleport out.
We could also make Abra's teleportation destination random, that would be fun. Along with the obligation to bring ONLY Koffing and Gastly, so the player could end up in difficult situations.