[feature] Sabrina's AI improvement (see details)
Note from previous commit:
Chuck can't switch out.
End of note.

A Pokémon that has both a trapping/binding move and perish song should do its combo in the right order: first the trapping move, then perish song, so it's too late for the player to switch out when he hears the perish song.

This should also prevent the AI from switching out the next turn after using Perish Song, as there is a case that makes the AI stay in until the last moment if the player is trapped/binded.