[feature] Erika's pollen added for her challenge (see details)! With a nice animation and stuff! I've outdone myself on this one, even though it's dirty work.
The pollen was supposed to work like the ability "Effect Spore" but with only the Sleep effect, and 100% chance on a contact move.
As we don't have the Contact info on moves, I decided to consider all Physical moves as contact moves.
Because I'm lazy, and don't want to make the project bigger, I plugged my function to the BattleCommand_HeldFlinch (kingsrock macro).
However I then discovered that a lot of physical moves (some considered contact moves in later generations) do not proc kingsrock.
The solution I opted for was to add the kingsrock macro to a lot of moves, breaking some of the game's rules, most notably the FLINCH_HIT moves that now have more odds of making the target flinch. Perhaps I could write a fix in the future to avoid this specific case.
The cleaner way to fix it would be to add the Contact info to my moves, but that's a lot of work but brings almost nothing to the gameplay (just 1 battle) and makes the UI heavier anyway.