[feature] The Patrol movedata is now working better! See notes:
However, the NPC can detect you while out of the screen: this should be fixed in the detection script (check if out of bounds).
Also, after a battle or a reload, the NPC considers it position as its origin position, and starts moving out of its range. Hence the need to remove roaming officers after each battle. Perhaps a simple line of code could fix this issue.