[feature] (see notes) the OW sprites of the police are now updated correctly!
The refresh is functional, but not implemented yet.
A call to either "disappear LAST_TALKED" or "reloadend MAPSETUP_CONNECTION" must be done to refresh the sprites.
If this could be done automatically, rather than on every map, that would be great.

In the same fashion, we could refresh the wCurFreedomState after every battle, so that the police comes as soon as possible and forbid the player from farming an area while he shouldn't be able to. Note that this could cause issues, as the player could stand above an NPC, which would prevent it from spawning and could break the game. Only refresh the map when the freedom state makes the police appear, but not the other way around could be a solution. Also, don't make key NPCs ever disappear.