[fix] Random memory jumps due to shore wave animation fixed.
The memory jumps were due to the Multiply function that use push and pop, which messed with the Stack Pointer (sp register), as the multiplication was done before we save sp.

Also, the scanline is still there. As a workaround, I'm stopping the animation when the player is in an area where it's not needed. I currently don't know what causes it. My guess is it's the frame delays cause by the writing in VRAM. Probably that the sprite is being written while VRAM is not writable.

Tomorrow, I'd like to try to write all 12 shore wave tiles in 1 pass once again. I initially split this function, as I though it was the cause of the glitches and bugs.