[feature] [BUGGED] smooth time of day pal transition is giving wrong colors
It's obvious, especially on the 4th color of each palette, which is 7,7,7 and shouldn't change.
Doesn't take the proper time of day.
The time of day offset is broken, as it's hardcoded as +8, which doesn't work from nite to morn (-8).
Also, it won't work as an animation in the long run.