[WIP] WIP: Emergency system.
Making good progress.
Next steps: Release from the hospital box when it is full, give a abandonned pkmn to the player if he has no other pkmn (or only eggs), determine the right condition for the nurse to take a pokémon and code it, create all the texts and animation for the nurse when it happens.