[feature] Mart's inventory selection affected by shortages. See notes.
Prices of items not affected yet.
Quantity of items buyable not capped yet.

The herb shop is fine (I can still change the prices if I want).
The pharmacy is also fine, and will have its prices updated.
The bargain shop is fine as it is, as long as I don't change the price of the items it sells.
The rooftop sale will need to be adjusted later.