[feature] very slightly optimized the font loading. See notes.
The result is bad in proportion of the amount of work involved.
I save a few palette loadings while navigating within the start menu, and 1 palette loading while using Fly.
It's not super efficient, and may cause bugs, so I should get rid of the commit.

A very efficient way to load the font, would be to have a background on the font (like the overworld font), and have the palette define the second color as white. And load the overworld_space in VRAM 0 address $7f. This way, we would almost never reload the palettes.